Fix string format warnings not being enabled in debug mode anymore
The check for the debug build type was always causing the `-Wno-format` flag to be added instead of the desired `-Wformat=2` flag in debug mode, as the build type is a string `Debug` not `DEBUG`.

+++ src/game/client/components/menus_ingame_touch_controls.cpp
@@ -88,7 +88,7 @@ void CMenusIngameTouchControls::RenderTouchButtonEditor(CUIRect MainView)
 	case EElementType::LAYOUT: Changed |= RenderLayoutSettingBlock(Block); break;
 	case EElementType::VISIBILITY: Changed |= RenderVisibilitySettingBlock(Block); break;
 	case EElementType::BEHAVIOR: Changed |= RenderBehaviorSettingBlock(Block); break;
-	default: dbg_assert(false, "Unknown m_EditElement = %d.", m_EditElement);
+	default: dbg_assert(false, "Unknown m_EditElement = %d.", (int)m_EditElement);
 	}
 
 	// Save & Cancel & Hint.
